setup.sh: Wait for successful HTTP status code, instead of any response
Now that we have a proxy in front of Prosody, it returns 502 when Prosody
is not ready yet, whereas we used to get no response.

By default curl still exits with a success code when there is any kind of
response, but -f (--fail) changes that and makes the 502 error-ish, which
should trigger the retry logic.